Move comparison functions registrar to runtime/standard.
PiperOrigin-RevId: 527917420
