+++ eval/public/BUILD
@@ -325,16 +325,10 @@ cc_library(
     deps = [
         ":cel_function_registry",
         ":cel_options",
-        "//base:builtins",
-        "//base:function_adapter",
-        "//base:handle",
-        "//base:value",
-        "//internal:status_macros",
         "//runtime:function_registry",
         "//runtime:runtime_options",
-        "//runtime/internal:number",
+        "//runtime/standard:comparison_functions",
         "@com_google_absl//absl/status",
-        "@com_google_absl//absl/time",
     ],
 )
 
@@ -346,29 +340,19 @@ cc_test(
     ],
     deps = [
         ":activation",
-        ":cel_builtins",
         ":cel_expr_builder_factory",
         ":cel_expression",
         ":cel_function_registry",
         ":cel_options",
         ":cel_value",
         ":comparison_functions",
-        ":message_wrapper",
-        "//eval/public/containers:container_backed_list_impl",
-        "//eval/public/containers:container_backed_map_impl",
-        "//eval/public/structs:cel_proto_wrapper",
-        "//eval/public/structs:trivial_legacy_type_info",
         "//eval/public/testing:matchers",
-        "//eval/testutil:test_message_cc_proto",
         "//internal:status_macros",
         "//internal:testing",
         "//parser",
-        "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/time",
-        "@com_google_absl//absl/types:span",
-        "@com_google_absl//absl/types:variant",
         "@com_google_googleapis//google/api/expr/v1alpha1:syntax_cc_proto",
         "@com_google_googleapis//google/rpc/context:attribute_context_cc_proto",
         "@com_google_protobuf//:protobuf",
